Fixed server crashing, for loop in purchased items was covering entire block so res.send was being sent for every loop
