+++ client/src/views/ShopView.js
@@ -2,10 +2,10 @@
 
 import React, { useState, useContext } from "react";
 import { Row, Col } from "react-bootstrap";
-import NavbarShop from '../components/NavbarShop';
 import ProductCard from "../components/ProductCard";
 import Search from "../components/Search";
 import ProductContext from "../ProductContext";
+import NavbarShop from "../components/NavbarShop";
 
 function ShopView(props) {
   const { searched } = useContext(ProductContext);
@@ -14,7 +14,7 @@ function ShopView(props) {
 
   return (
     <>
-    <NavbarShop></NavbarShop>
+    {/* <NavbarShop></NavbarShop> */}
       <h1>Welcome to the store!</h1>
       <Search />
       <Row xs={1} md={3} className="g-4">
